Completed POJO, Alarm Creator set beautifully.
Yet to come, Edit Alarm Activity.

Post which, setting the alarm service in android background. Making sure it stays beyond reboot.

+++ app/src/main/java/com/theflopguyproductions/ticktrack/ui/counter/activity/counter/CounterActivity.java
@@ -53,7 +53,7 @@ public class CounterActivity extends AppCompatActivity {
         counterDataArrayList = gson.fromJson(json, type);
 
 
-        currentPosition = (int) getIntent().getIntExtra("currentPosition",0);
+        currentPosition = getIntent().getIntExtra("currentPosition",0);
         currentCount = counterDataArrayList.get(currentPosition).getCountValue();
         counterLabel.setText(counterDataArrayList.get(currentPosition).getCounterLabel());
